# swagger.yaml
consumes:
- application/json
definitions:
  Product:
    properties:
      description:
        description: The description for this product
        type: string
        x-go-name: Description
      id:
        description: The product id
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: The name for this product
        type: string
        x-go-name: Name
      price:
        description: The price for this product
        format: float
        minimum: 0
        type: number
        x-go-name: Price
      sku:
        description: The SKU for this product
        type: string
        x-go-name: SKU
    required:
    - id
    - name
    - price
    type: object
    x-go-package: github.com/mserebryaakov/microservices-training/data
host: localhost
info:
  description: Documentation for Product API
  title: of Product API.
  version: 1.0.0
paths:
  /products:
    get:
      description: Return a list of products from the database
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: Create product with the enter id from the database
      operationId: createProduct
      parameters:
      - $ref: '#/definitions/Product'
        description: Product in the system
        in: path
        name: Body
        required: true
      responses:
        "200":
          $ref: '#/responses/noContent'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete product with the enter id from the database
      operationId: deleteProduct
      parameters:
      - description: The id of the product to delete from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
        "400":
          $ref: '#/responses/errorResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    put:
      description: Update product with the enter id from the database
      operationId: updateProduct
      parameters:
      - $ref: '#/definitions/Product'
        description: Product in the system
        in: path
        name: Body
        required: true
      responses:
        "201":
          $ref: '#/responses/noContent'
        "400":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - products
produces:
- application/json
responses:
  NoContent:
    description: ""
  errorResponse:
    description: Response ERROR
  errorValidation:
    description: Validation ERROR
  productsResponse:
    description: A list of products returns in the response
    headers:
      Body:
        description: |-
          All products in the system
          in:Body
        items:
          $ref: '#/definitions/Product'
        type: array
schemes:
- http
swagger: "2.0"